backend/requirements.txt
asgiref==3.11.1
drf-spectacular==0.28.0
whitenoise==6.9.0
django-jazzmin==3.0.4
django-redis==5.4.0
pytest==8.3.5
pytest-django==4.10.0
async-timeout==5.0.1
attrs==26.1.0
autobahn==24.4.2
Automat==25.4.16
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
packaging==26.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
pyOpenSSL==26.0.0
redis==7.4.0
service-identity==24.2.0
sqlparse==0.5.5
tomli==2.4.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
zope.interface==8.2
